+++ src/com/example/views/concrete/EllipticStyle.java
@@ -7,12 +7,12 @@ import java.awt.geom.Ellipse2D;
 
 public class EllipticStyle extends Style {
 
-    public EllipticStyle(Color newColor, double newWidth, double newHeight) {
+    public EllipticStyle(Color newColor, int newWidth, int newHeight) {
         super(newColor, newWidth, newHeight);
     }
 
     @Override
-    public Shape makeshape(double x, double y, double width, double height) {
+    public Shape makeshape(int x, int y, int width, int height) {
         return new Ellipse2D.Double(x, y, width, height);
     }
 }
